[MISC, TEST] Rename header and test files
