+++ test/snippet/alphabet/quality/phred68solexa.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/alphabet/quality/phred68legacy.hpp>
+#include <seqan3/alphabet/quality/phred68solexa.hpp>
 #include <seqan3/core/debug_stream.hpp>
 
 int main()
