+++ test/snippet/alphabet/quality/phred68solexa_literal.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/alphabet/quality/phred68legacy.hpp>
+#include <seqan3/alphabet/quality/phred68solexa.hpp>
 #include <seqan3/core/debug_stream.hpp>
 #include <seqan3/std/algorithm>
