+++ test/unit/alphabet/quality/CMakeLists.txt
@@ -1,7 +1,7 @@
 
 seqan3_test(phred42_test.cpp)
 seqan3_test(phred63_test.cpp)
-seqan3_test(phred68legacy_test.cpp)
+seqan3_test(phred68solexa_test.cpp)
 seqan3_test(phred94_test.cpp)
 seqan3_test(qualified_test.cpp)
 seqan3_test(quality_conversion_integration_test.cpp)
